app: improved algorithm for action search.
Now matches have various levels of reliability.
If the keyword is 2 characters, the match on first letters of first 2
words is most reliable; then a label prefix, then a substring anywhere
in the label; then a substring anywhere in the tooltip; and finally a
fuzzy match. This new algorithm would return the same list as before,
but in a much more significative order.
I also got rid of the " " = show all trick, replaced by a NULL keyword.